+++ test/socketio/pcp_socket.io.test.js
@@ -71,8 +71,8 @@ const pcpTriggerSideEffects = `pcp-action:triggerSideEffects
 
 const pcpSideEffect1 = `pcp-action:MESSAGE
 pcp-channel:amc\\://notification/notify
-sideEffectsSource:/Header(ID='e0582b6a-6d93-46d9-bd28-98723a285d40')
-sideEffectsEventName:sideEffect1
+sideEffectSource:/Header(ID='e0582b6a-6d93-46d9-bd28-98723a285d40')
+sideEffectEventName:sideEffect1
 serverAction:RaiseSideEffect
 
 `;